feat(practica03): :construction: Antonio Gil · añadir tests del DAO anticorrupción

+++ AntonioGil/practica03/src/main/java/com/viewnext/bsan/practica03/dao/impl/UserDaoImpl.java
@@ -75,7 +75,9 @@ public class UserDaoImpl implements UserDao {
     @Override
     public void deleteByDni(String dni) throws DaoLevelException {
         try {
-            repository.deleteById(dni);
+            if (dni != null) {
+                repository.deleteById(dni);
+            }
         } catch (NestedRuntimeException ex) {
             throw new DaoLevelException("UserDao#deleteByDni failed", ex);
         }
